+++ pkg/config/config_test.go
@@ -16,6 +16,8 @@ func createEnvFile(t *testing.T) func() {
 
 	_, err = f.WriteString("ETH_PRIVATE_KEY=\"0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef\"")
 	testutil.Ok(t, err)
+	_, err = f.WriteString("NODE_URL=\"https://mainnet.infura.io/v3/xxxxxxxxxxxxx\"")
+	testutil.Ok(t, err)
 	testutil.Ok(t, f.Close())
 
 	return func() {